Make a top level full app that doesn't render itself.
