+++ package.json
@@ -17,15 +17,13 @@
         "@emotion/css": "^11.1.3",
         "@emotion/react": "^11.5.0",
         "@ironcorelabs/ironweb": "^3.4.1",
-        "es6-promise": "^4.2.8",
         "jsonwebtoken": "^8.5.1",
         "react": "17.0.2",
         "react-dom": "17.0.2",
         "react-redux": "^7.2.4",
         "react-scripts": "4.0.3",
         "redux": "^4.1.0",
-        "redux-thunk": "^2.3.0",
-        "whatwg-fetch": "^3.6.2"
+        "redux-thunk": "^2.3.0"
     },
     "prettier": {
         "printWidth": 160,
